requirements.txt
torch>=2.0.0
open_clip_torch>=2.24.0
Pillow>=10.0.0
pyyaml>=6.0.0
tqdm>=4.65.0
numpy>=1.24.0
